--- apps/web/src/app/[locale]/shop/SearchOption/index.ts
@@ -1,4 +1,0 @@
-import OrderTypeSelect from './OrderTypeSelect';
-import SortDirectionSelect from './SortDirectionSelect';
-
-export { OrderTypeSelect, SortDirectionSelect };
